Refactor VOC Augmenter to support additional scaling method
The `VOCAuger` class in `voc2012_auger2.py` has been modified to include support for a new scaling method called 'fix_short_no_crop'. This method allows the user to specify a target size and an optional aligner function. If an aligner function is provided, it will be used during the scaling process. Otherwise, the default aligner function will be used.

This change improves the flexibility of the VOC Augmenter by providing an alternative scaling option for users who do not want to perform cropping during augmentation.